+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Removed UnSync from dialog,
 - Updated private/public msg
+- Skipping Update .syncignore step
 ### Added
 - Normalizing file paths for windows